+++ packages/native/src/Radio/types.ts
@@ -8,13 +8,13 @@ export type NumberStringValue = string | number;
 
 export interface RadioProps {
   activeColor?: ColorType;
-  deactiveColor?: ColorType;
+  inactiveColor?: ColorType;
+  isActive?: boolean;
   label: string;
+  onPress?: (value: NumberStringValue) => void;
   size?: SizeType;
   type?: RadioType;
   value: NumberStringValue;
-  onPress?: (value: NumberStringValue) => void;
-  isActive?: boolean;
 }
 
 type AlignType = 'horizontal' | 'vertical';
@@ -23,7 +23,7 @@ export interface RadioGroupProps {
   activeColor?: ColorType;
   align?: AlignType;
   children?: ReactElement<RadioProps> | ReactElement<RadioProps>[];
-  deactiveColor?: ColorType;
+  inactiveColor?: ColorType;
   defaultValue?: NumberStringValue;
   error?: boolean;
   onChange?: (key: NumberStringValue) => void;